Revert "Refactor pipeline build (#75)"
This reverts commit 6b9bb721cb6d7f009045b0c9bd2f391f08bf45cb.
