+++ src/spdl/dataloader/__init__.py
@@ -5,7 +5,6 @@
 from . import (  # noqa: E402
     _bg_consumer,
     _bg_generator,
-    _builder,
     _flist,
     _hook,
     _pipeline,
@@ -15,7 +14,6 @@ from . import (  # noqa: E402
 _mods = [
     _bg_consumer,
     _bg_generator,
-    _builder,
     _hook,
     _pipeline,
     _utils,
